+++ source/renderers/sdl3/image_sdl3.cpp
@@ -13,8 +13,8 @@
 #include <vector>
 
 void Image_SDL3::render(ImageRenderParams &params) {
-    const int &x = params.x;
-    const int &y = params.y;
+    const float &x = params.x;
+    const float &y = params.y;
     const int &brightness = params.brightness;
     const double rotation = Math::radiansToDegrees(params.rotation);
     const float &scale = params.scale;
